fix: fold workers scheduler constructor into poller_watcher for pkg-file-count
